+++ src/main/kotlin/com/github/encryptsl/magenta/listeners/EntityListeners.kt
@@ -17,6 +17,7 @@ import org.bukkit.event.entity.EntityDamageByEntityEvent
 import org.bukkit.event.entity.EntityDamageEvent
 import org.bukkit.inventory.EquipmentSlot
 import org.bukkit.inventory.ItemStack
+import java.math.BigDecimal
 import java.util.concurrent.ThreadLocalRandom
 
 class EntityListeners(private val magenta: Magenta) : HalloweenAPI(), Listener {
@@ -55,7 +56,7 @@ class EntityListeners(private val magenta: Magenta) : HalloweenAPI(), Listener {
             if (!magenta.config.contains("jobs.hunter") || !magenta.config.contains("jobs.hunter.earn_money")) return
 
             val earnMoney = magenta.config.getDouble("halloween.reward_multiplier").times(magenta.config.getDouble("jobs.hunter.earn_money", 5.0))
-            val transaction = EconomyDeposit(player, earnMoney).transaction(magenta.vaultHook) ?: return
+            val transaction = EconomyDeposit(player, BigDecimal.valueOf(earnMoney)).transaction(magenta.vaultHook) ?: return
 
             if (transaction == EconomyTransactionResponse.SUCCESS) {
                 player.sendActionBar(ModernText.miniModernText(magenta.config.getString("jobs.hunter.earn_bar").toString(),
